Code/traffic-predition/requirements.txt
Django>=5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
gunicorn>=21.2.0
numpy>=1.26.4
pandas>=2.2.2
joblib>=1.3.2
scikit-learn>=1.4.2
requests>=2.31.0
python-decouple>=3.8  # Optional, for environment variables
